# API-Specification.json
{
    "swagger": "2.0",
    "info": {
        "description": "API For Customer PhoneNumbers",
        "version": "1.0",
        "title": "Customer PhoneNumber API",
        "termsOfService": "Terms of Service",
        "contact": {
            "name": "Rajni Saluja",
            "url": "www.test.com",
            "email": "rajni.saluja@abc.com"
        },
        "license": {
            "name": "License of API",
            "url": "API license URL"
        }
    },
    "host": "localhost:8085",
    "basePath": "/",
    "tags": [
        {
            "name": "customer-controller",
            "description": "Customer Controller"
        }
    ],
    "paths": {
        "/customers/phonenumbers": {
            "get": {
                "tags": [
                    "customer-controller"
                ],
                "summary": "Get All Customer's PhoneNumbers",
                "operationId": "getAllCustPhoneNumbersUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "customerIdToFetchFrom",
                        "in": "query",
                        "description": "customerIdToFetchFrom",
                        "required": false,
                        "type": "integer",
                        "default": 1,
                        "minimum": 1,
                        "exclusiveMinimum": false,
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CustomerList"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/customers/{id}/phonenumbers": {
            "get": {
                "tags": [
                    "customer-controller"
                ],
                "summary": "Get PhoneNumbers of a particular customer",
                "operationId": "getPhoneNumbersByIdUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "integer",
                        "minimum": 1,
                        "exclusiveMinimum": false,
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Customer"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "customer-controller"
                ],
                "summary": "Activate all/particular phonenumber of a customer",
                "operationId": "activatePhoneNumberUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "integer",
                        "minimum": 1,
                        "exclusiveMinimum": false,
                        "format": "int32"
                    },
                    {
                        "in": "body",
                        "name": "number",
                        "description": "number",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ActiveNumber"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Customer"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "ActiveNumber": {
            "type": "object",
            "required": [
                "activate"
            ],
            "properties": {
                "activate": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            },
            "title": "ActiveNumber"
        },
        "Customer": {
            "type": "object",
            "properties": {
                "customerId": {
                    "type": "integer",
                    "format": "int32"
                },
                "phoneNumbers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/PhoneNumber"
                    }
                }
            },
            "title": "Customer"
        },
        "CustomerList": {
            "type": "object",
            "properties": {
                "customers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Customer"
                    }
                },
                "nextCustomerToFetch": {
                    "type": "integer",
                    "format": "int32"
                }
            },
            "title": "CustomerList"
        },
        "PhoneNumber": {
            "type": "object",
            "properties": {
                "activate": {
                    "type": "boolean"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            },
            "title": "PhoneNumber"
        }
    }
}